requirements.txt
fastapi[all]
sqlalchemy
python-dotenv
pandas
numpy
plotly
uvicorn
psycopg2